moves copy functionality to Copiable component
