+++ src/storyHelpers/ColorTokenSwatchGroup.svelte
@@ -1,6 +1,5 @@
 <script>
-  import Label from '../components/label/label.svelte'
-  import Icon from '../components/icon/icon.svelte'
+  import Copiable from './Copiable.svelte'
 
   export let level = 1
   export let name = ''
@@ -11,42 +10,25 @@
   const getTokens = (obj) =>
     Object.entries(obj).filter((i) => i[0] !== 'toString')
 
-  const handleClick = async ({ currentTarget }) => {
-    await navigator.clipboard.writeText(currentTarget.title)
-    currentTarget.classList.add('isCopied')
-    setTimeout(() => {
-      currentTarget.classList.remove('isCopied')
-    }, 1000)
-  }
-
   const filteredTokens = getTokens(tokens)
 </script>
 
 <svelte:element this={name ? 'section' : 'div'} id={name}>
   {#if name}
-    <svelte:element this={headingTag} class="group-heading"
-      >{name}</svelte:element
-    >
+    <svelte:element this={headingTag} class="group-heading">
+      {name}
+    </svelte:element>
   {/if}
 
   {#if typeof filteredTokens[0][1] === 'string'}
     <div class="color-swatch-group">
       {#each filteredTokens as [name, token]}
-        <button
-          title={token}
-          class="color-swatch"
-          style="--swatch-bg:{token}"
-          on:click={handleClick}
-        >
-          <span class="copied">
-            <Label color="green" mode="loud">
-              <Icon name="check-circle-filled" />
-              Copied!
-            </Label>
-          </span>
-          <span class="sample" />
-          {name}
-        </button>
+        <Copiable text={token}>
+          <div class="color-swatch" style="--swatch-bg:{token}">
+            <span class="sample" />
+            {name}
+          </div>
+        </Copiable>
       {/each}
     </div>
   {:else}
@@ -92,16 +74,5 @@
     width: 100%;
     aspect-ratio: 1/1;
     background: var(--swatch-bg);
-  }
-
-  .color-swatch .copied {
-    position: absolute;
-    top: 30%;
-    opacity: 0;
-    transition: opacity 0.13s ease-out;
-  }
-
-  .color-swatch:global(.isCopied) .copied {
-    opacity: 1;
   }
 </style>
