+++ src/storyHelpers/FontTokenSwatchGroup.svelte
@@ -1,6 +1,5 @@
 <script>
-  import Label from '../components/label/label.svelte'
-  import Icon from '../components/icon/icon.svelte'
+  import Copiable from "./Copiable.svelte"
 
   export let level = 1
   export let name = ''
@@ -11,22 +10,14 @@
   const getTokens = (obj) =>
     Object.entries(obj).filter((i) => i[0] !== 'toString')
 
-  const handleClick = async ({ currentTarget }) => {
-    await navigator.clipboard.writeText(currentTarget.title)
-    currentTarget.classList.add('isCopied')
-    setTimeout(() => {
-      currentTarget.classList.remove('isCopied')
-    }, 1000)
-  }
-
   const filteredTokens = getTokens(tokens)
 </script>
 
 <svelte:element this={name ? 'section' : 'div'} id={name}>
   {#if name}
-    <svelte:element this={headingTag} class="group-heading"
-      >{name}</svelte:element
-    >
+    <svelte:element this={headingTag} class="group-heading">
+      {name}
+    </svelte:element>
   {/if}
 
   {#if typeof filteredTokens[0][1] === 'string'}
@@ -35,20 +26,9 @@
         <svelte:element this={`h${level + 1}`} class="group-heading"
           >{name}</svelte:element
         >
-        <button
-          title={token}
-          class="font"
-          style="--swatch-bg:{token}"
-          on:click={handleClick}
-        >
-          <span class="copied">
-            <Label color="green" mode="loud">
-              <Icon name="check-circle-filled" />
-              Copied!
-            </Label>
-          </span>
-          <span class="sample" style="font: {token}">Keep it bravey.</span>
-        </button>
+        <Copiable text={token}>
+          <span class="font-sample" style="font: {token}">Keep it bravey.</span>
+        </Copiable>
       {/each}
     </div>
   {:else}
@@ -75,33 +55,11 @@
     padding-bottom: 1rem;
   }
 
-  .font {
+  .font-sample {
     display: flex;
     font-size: 0.75rem;
     color: var(--leo-color-text-primary);
-    gap: 0.5rem;
-    margin: 0;
-    padding: 0;
-    border: none;
-    background: none;
-    cursor: pointer;
-    position: relative;
-  }
-
-  .font .sample {
     background: var(--leo-color-page-background);
-    width: 100%;
     padding: 2rem;
-  }
-
-  .font .copied {
-    position: absolute;
-    right: 30%;
-    opacity: 0;
-    transition: opacity 0.13s ease-out;
-  }
-
-  .font:global(.isCopied) .copied {
-    opacity: 1;
   }
 </style>
